Changed models
Added user role attribute in User class and removed user role models. Changed ID type from Integer to Long for some models. Replaced javax import with jakarta.

+++ src/main/java/com/ISS/Booking_iss_tim21/repository/AccommodationReviewRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface AccommodationReviewRepository extends JpaRepository<AccommodationReview,Integer> {
+public interface AccommodationReviewRepository extends JpaRepository<AccommodationReview,Long> {
 }
