+++ src/main/java/com/ISS/Booking_iss_tim21/repository/ReviewReportRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ReviewReportRepository extends JpaRepository<ReviewReport,Integer> {
+public interface ReviewReportRepository extends JpaRepository<ReviewReport,Long> {
 }
